+++ config/webpack.config.dev.js
@@ -224,7 +224,9 @@ module.exports = {
                       '@entities': path.resolve(paths.appLibSrc, './entities'),
                       '@utils': path.resolve(paths.appLibSrc, './utils'),
                       '@hooks': path.resolve(paths.appLibSrc, './hooks'),
-                      '@context': path.resolve(paths.appLibSrc, './context')
+                      '@context': path.resolve(paths.appLibSrc, './context'),
+                      '@classes': path.resolve(__dirname, '../src/lib/classes'),
+                      '@constants': path.resolve(__dirname, '../src/lib/constants')
                     }
                   }
                 ]