+++ src/lib/classes/index.js
@@ -1,0 +1,5 @@
+import { Notification } from './notification';
+import AccessControlList from './access-control-list';
+import AppPermission from './app-permissions';
+
+export { Notification, AccessControlList, AppPermission };